+++ context_response.go
@@ -27,22 +27,6 @@
 
 package iris
 
-type (
-	// IContextResponse is part of the IContext
-	IContextResponse interface {
-		// SetContentType sets the "Content-Type" header, receives the values
-		SetContentType([]string)
-		// SetHeader sets the response headers first parameter is the key, second is the values
-		SetHeader(string, []string)
-		Redirect(string, ...int)
-		// Errors
-		NotFound()
-		Panic()
-		EmitError(int)
-		//
-	}
-)
-
 // SetContentType sets the response writer's header key 'Content-Type' to a given value(s)
 func (ctx *Context) SetContentType(s []string) {
 	for _, hv := range s {