+++ iris.go
@@ -25,7 +25,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-// Package iris v2.2.4
+// Package iris v2.3.0
 //
 // Note: When 'Station', we mean the Iris type.
 package iris
@@ -34,16 +34,42 @@ import (
 	"os"
 
 	"sync"
+	"time"
 
 	"github.com/kataras/iris/logger"
 	"github.com/kataras/iris/render"
 	"github.com/kataras/iris/server"
+	"github.com/kataras/iris/sessions"
+	_ "github.com/kataras/iris/sessions/providers/memory"
+	_ "github.com/kataras/iris/sessions/providers/redis"
+)
+
+const (
+	// TemplateHTML html/template
+	TemplateHTML TemplateEngine = iota
+	//when v3 add:  TemplatePongo2
 )
 
 type (
+	TemplateEngine uint8
+
+	// SessionConfig the configuration for sessions
+	// We don't import the providers and make it easier with Provider = iris.Redis OR iris.Memory [iotas] and make all the rest automatically because
+	// we want to give the developers the functionality to change the options of each now/and future/or custom session provider they select
+	// example: import "github.com/kataras/iris/sessions/providers/redis" ... redis.Config.Addr = "127.0.0.1:2222";iris.Config().Session.Provider = redis.Provider
+	SessionConfig struct {
+		// Provider string, usage iris.Config().Provider = "memory" or "redis". If you wan to customize redis then import the package, and change it's config
+		Provider string
+		// Secret string, the session's client cookie name, for example: "irissessionid"
+		Secret string
+		// Life time.Duration, cookie life duration and gc duration, for example: time.Duration(60)*time.Minute
+		Life time.Duration
+	}
+
 	// IrisConfig options for iris before server listen
 	// MaxRequestBodySize is the only options that can be changed after server listen - using SetMaxRequestBodySize(int)
-	// Render can be changed after declaration but before server's listen - using SetRenderConfig(RenderConfig)
+	// Render config can be changed after declaration but before server's listen - using Config().Render
+	// Session config can be changed after declaration but before server's listen - using Config().Session
 	IrisConfig struct {
 		// MaxRequestBodySize Maximum request body size.
 		//
@@ -78,18 +104,31 @@ type (
 		// Default is /debug/pprof , which means yourhost.com/debug/pprof
 		ProfilePath string
 
-		// Render specify configs for rendering
-		Render *RenderConfig
+		// TemplateEngine the engine for rendering templates [No usage yet, wait for iris-v3]
+		TemplateEngine TemplateEngine
+		// Render configs for rendering.
+		// Render has some options for rendering with html/template only, we will keep this as it is on iris-v3 too.
+		//
+		// these options inside this config don't have any relation with the TemplateEngine
+		// from github.com/kataras/iris/render
+		Render *render.Config
+
+		// Session the config for sessions
+		// contains 3(three) properties
+		// Provider: (look /sessions/providers)
+		// Secret: cookie's name (string)
+		// Life: cookie life (time.Duration)
+		Session *SessionConfig
 	}
 
 	// Iris is the container of all, server, router, cache and the sync.Pool
 	Iris struct {
 		*router
-		server  *server.Server
-		plugins *PluginContainer
-		render  *render.Render
-		//we want options exported, Options but Options is an http method also, so we make a big change here
-		// and rename the iris.IrisOptions to simple 'iris.IrisConfig' - no iris.config because of the default func Config()
+		server         *server.Server
+		plugins        *PluginContainer
+		render         *render.Render
+		sessionManager *sessions.Manager
+
 		config *IrisConfig
 		logger *logger.Logger
 	}
@@ -119,6 +158,8 @@ func New(configs ...*IrisConfig) *Iris {
 	s.logger = logger.New()
 	s.logger.SetEnable(config.Log)
 
+	// set the render (no build templates, to give the chance to the users to change its *Config)
+	s.render = render.Create(config.Render)
 	return s
 }
 
@@ -142,6 +183,11 @@ func (s *Iris) Logger() *logger.Logger {
 	return s.logger
 }
 
+// Render returns the render
+func (s *Iris) Render() *render.Render {
+	return s.render
+}
+
 // SetMaxRequestBodySize Maximum request body size.
 //
 // The server rejects requests with bodies exceeding this limit.
@@ -149,11 +195,6 @@ func (s *Iris) Logger() *logger.Logger {
 // By default request body size is unlimited.
 func (s *Iris) SetMaxRequestBodySize(size int) {
 	s.config.MaxRequestBodySize = size
-}
-
-// SetRenderConfig sets the Config.Render, can be setted before server's listen, not after.
-func (s *Iris) SetRenderConfig(renderCfg *RenderConfig) {
-	s.config.Render = renderCfg
 }
 
 // newContextPool returns a new context pool, internal method used in tree and router
@@ -169,6 +210,15 @@ func (s *Iris) newContextPool() sync.Pool {
 // it's a non-blocking func
 func (s *Iris) DoPreListen(opt server.Config) *server.Server {
 	//runs only once even if called more than one time.
+
+	// build/prepare the  render now.
+	if s.render == nil { // if it's nil ( that's not happening normally, it is setted on .New() )
+		s.render = render.New(s.config.Render)
+	}
+
+	s.render.Prepare()
+
+	// router prepare
 	if !s.router.optimized {
 		s.router.optimize()
 
@@ -178,9 +228,9 @@ func (s *Iris) DoPreListen(opt server.Config) *server.Server {
 		if s.config.MaxRequestBodySize > 0 {
 			s.server.MaxRequestBodySize = s.config.MaxRequestBodySize
 		}
-
-		s.plugins.DoPreListen(s)
 	}
+
+	s.plugins.DoPreListen(s)
 
 	return s.server
 }
@@ -189,12 +239,17 @@ func (s *Iris) DoPreListen(opt server.Config) *server.Server {
 // it's a non-blocking func
 func (s *Iris) DoPostListen() {
 
-	if s.render == nil {
-		// set the render(er) now
-		s.render = newRender(s.config.Render)
-		s.plugins.DoPostListen(s)
+	if s.config.Session != nil && s.config.Session.Provider != "" {
+		if s.config.Session.Secret == "" {
+			s.config.Session.Secret = DefaultCookieName
+		}
+		if s.config.Session.Life == 0 {
+			s.config.Session.Life = DefaultCookieDuration
+		}
+		s.sessionManager = sessions.New(s.config.Session.Provider, s.config.Session.Secret, s.config.Session.Life)
 	}
 
+	s.plugins.DoPostListen(s)
 }
 
 // openServer is internal method, open the server with specific options passed by the Listen and ListenTLS