+++ sessions/provider.go
@@ -31,12 +31,15 @@ import (
 	"container/list"
 	"sync"
 	"time"
+
+	"github.com/kataras/iris/sessions/store"
 )
 
 // IProvider the type which Provider must implement
 type IProvider interface {
-	Init(string) (IStore, error)
-	Read(string) (IStore, error)
+	Name() string
+	Init(string) (store.IStore, error)
+	Read(string) (store.IStore, error)
 	Destroy(string) error
 	Update(string) error
 	GC(time.Duration)
@@ -46,11 +49,12 @@ type (
 	// Provider implements the IProvider
 	// contains the temp sessions memory, the store and some options for the cookies
 	Provider struct {
+		name               string
 		mu                 sync.Mutex
 		sessions           map[string]*list.Element // underline TEMPORARY memory store
 		list               *list.List               // for GC
-		NewStore           func(sessionId string, cookieLifeDuration time.Duration) IStore
-		OnDestroy          func(store IStore) // this is called when .Destroy
+		NewStore           func(sessionId string, cookieLifeDuration time.Duration) store.IStore
+		OnDestroy          func(store store.IStore) // this is called when .Destroy
 		cookieLifeDuration time.Duration
 	}
 )
@@ -58,14 +62,14 @@ type (
 var _ IProvider = &Provider{}
 
 // NewProvider returns a new empty Provider
-func NewProvider() *Provider {
-	provider := &Provider{list: list.New()}
+func NewProvider(name string) *Provider {
+	provider := &Provider{name: name, list: list.New()}
 	provider.sessions = make(map[string]*list.Element, 0)
 	return provider
 }
 
 // Init creates the store for the first time for this session and returns it
-func (p *Provider) Init(sid string) (IStore, error) {
+func (p *Provider) Init(sid string) (store.IStore, error) {
 	p.mu.Lock()
 
 	newSessionStore := p.NewStore(sid, p.cookieLifeDuration)
@@ -77,9 +81,9 @@ func (p *Provider) Init(sid string) (IStore, error) {
 }
 
 // Read returns the store which sid parameter is belongs
-func (p *Provider) Read(sid string) (IStore, error) {
+func (p *Provider) Read(sid string) (store.IStore, error) {
 	if elem, found := p.sessions[sid]; found {
-		return elem.Value.(IStore), nil
+		return elem.Value.(store.IStore), nil
 	}
 	// if not found
 	sessionStore, err := p.Init(sid)
@@ -90,7 +94,7 @@ func (p *Provider) Read(sid string) (IStore, error) {
 // Destroy always returns a nil error, for now.
 func (p *Provider) Destroy(sid string) error {
 	if elem, found := p.sessions[sid]; found {
-		elem.Value.(IStore).Destroy()
+		elem.Value.(store.IStore).Destroy()
 		delete(p.sessions, sid)
 		p.list.Remove(elem)
 	}
@@ -104,7 +108,7 @@ func (p *Provider) Update(sid string) error {
 	p.mu.Lock()
 
 	if elem, found := p.sessions[sid]; found {
-		elem.Value.(IStore).SetLastAccessedTime(time.Now())
+		elem.Value.(store.IStore).SetLastAccessedTime(time.Now())
 		p.list.MoveToFront(elem)
 	}
 
@@ -125,12 +129,17 @@ func (p *Provider) GC(duration time.Duration) {
 		}
 
 		// if the time has passed. session was expired, then delete the session and its memory place
-		if (elem.Value.(IStore).LastAccessedTime().Unix() + duration.Nanoseconds()) < time.Now().Unix() {
+		if (elem.Value.(store.IStore).LastAccessedTime().Unix() + duration.Nanoseconds()) < time.Now().Unix() {
 			p.list.Remove(elem)
-			delete(p.sessions, elem.Value.(IStore).ID())
+			delete(p.sessions, elem.Value.(store.IStore).ID())
 
 		} else {
 			break
 		}
 	}
+}
+
+// Name the provider's name, example: 'memory' or 'redis'
+func (p *Provider) Name() string {
+	return p.name
 }