+++ sessions/providers/redis/register.go
@@ -31,6 +31,8 @@ import (
 	"time"
 
 	"github.com/kataras/iris/sessions"
+	"github.com/kataras/iris/sessions/providers/redis/service"
+	"github.com/kataras/iris/sessions/store"
 )
 
 func init() {
@@ -38,9 +40,11 @@ func init() {
 }
 
 var (
-	provider = sessions.NewProvider()
-	// Redis is the empty redis service, you can set configs via this object
-	Redis = Empty()
+	Provider = sessions.NewProvider("redis")
+	// redis is the empty redis service, you can set configs via this object
+	redis = service.Empty()
+	// Config is just the Redis(service)' config
+	Config = redis.Config
 
 // Empty() because maybe the user wants to edit the default configs.
 //the Connect goes to the first NewStore, when user ask for session, so you have the time to change the default configs
@@ -50,11 +54,11 @@ var (
 // must runs only once
 func register() {
 	// the actual work is here.
-	provider.NewStore = func(sessionId string, cookieLifeDuration time.Duration) sessions.IStore {
+	Provider.NewStore = func(sessionId string, cookieLifeDuration time.Duration) store.IStore {
 		//println("memory.go:49-> requesting new memory store with sessionid: " + sessionId)
-		if !Redis.Connected {
-			Redis.Connect()
-			_, err := Redis.PingPong()
+		if !redis.Connected {
+			redis.Connect()
+			_, err := redis.PingPong()
 			if err != nil {
 				if err != nil {
 					// don't use to get the logger, just prin these to the console... atm
@@ -66,5 +70,5 @@ func register() {
 		return NewStore(sessionId, cookieLifeDuration)
 	}
 
-	sessions.Register("redis", provider)
+	sessions.Register(Provider)
 }