+++ sessions/store/store.go
@@ -25,7 +25,8 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-package sessions
+// Package store the package is in diffent folder to reduce the import cycles from the ./context/context.go *
+package store
 
 import "time"